refactor(web): split marketing navbar auth from Storybook view
Keep AuthKit out of the presentational navbar so Storybook can render without loading WorkOS client modules.
